Migrate monitoring to Slack proto
