+++ common/monitoring/build.gradle.kts
@@ -7,6 +7,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(projects.common.network)
+                implementation(projects.slack.proto)
             }
         }
     }
