--- common/monitoring/src/commonMain/kotlin/com/gchristov/thecodinglove/common/monitoring/slack/ApiSlackReportException.kt
@@ -1,10 +1,0 @@
-package com.gchristov.thecodinglove.common.monitoring.slack
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class ApiSlackReportException(
-    @SerialName("message") val message: String,
-    @SerialName("stacktrace") val stacktrace: String,
-)
