--- common/monitoring/src/commonMain/kotlin/com/gchristov/thecodinglove/common/monitoring/slack/SlackReportExceptionApi.kt
@@ -1,18 +1,0 @@
-package com.gchristov.thecodinglove.common.monitoring.slack
-
-import com.gchristov.thecodinglove.common.network.NetworkClient
-import io.ktor.client.request.*
-import io.ktor.client.statement.*
-import io.ktor.http.*
-
-internal class SlackReportExceptionApi(
-    private val client: NetworkClient.Json,
-    private val apiUrl: String,
-) {
-    suspend fun reportException(
-        exception: ApiSlackReportException
-    ): HttpResponse = client.http.post("$apiUrl/slack/report-exception") {
-        contentType(ContentType.Application.Json)
-        setBody(exception)
-    }
-}
