+++ search/adapter/src/commonMain/kotlin/com/gchristov/thecodinglove/search/adapter/db/DbSearchPost.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class DbSearchPost(
+internal data class DbSearchPost(
     @SerialName("title") val title: String,
     @SerialName("url") val url: String,
     @SerialName("image_url") val imageUrl: String,
