+++ search/service/build.gradle.kts
@@ -9,6 +9,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(projects.common.firebase)
+                implementation(projects.slack.proto)
                 implementation(projects.search.domain)
                 implementation(projects.search.adapter)
             }
