+++ search/service/src/jsMain/kotlin/com/gchristov/thecodinglove/search/service/SearchService.kt
@@ -21,6 +21,7 @@ import com.gchristov.thecodinglove.search.adapter.SearchAdapterModule
 import com.gchristov.thecodinglove.search.adapter.http.*
 import com.gchristov.thecodinglove.search.domain.SearchDomainModule
 import com.gchristov.thecodinglove.search.domain.model.Environment
+import com.gchristov.thecodinglove.slack.proto.SlackProtoModule
 
 suspend fun main() {
     // Ignore default Node arguments
@@ -48,8 +49,9 @@ private fun setupDi(environment: Environment): Either<Throwable, Unit> {
             CommonKotlinModule.module,
             CommonNetworkModule.module,
             CommonPubSubModule.module,
-            CommonMonitoringModule(environment.apiUrl).module,
+            CommonMonitoringModule.module,
             CommonFirebaseModule.module,
+            SlackProtoModule(environment.apiUrl).module,
             SearchAdapterModule(environment).module,
             SearchDomainModule.module,
         )
