+++ self-destruct/service/src/jsMain/kotlin/com/gchristov/thecodinglove/selfdestruct/service/SelfDestructService.kt
@@ -43,7 +43,7 @@ private fun setupDi(environment: Environment): Either<Throwable, Unit> {
         listOf(
             CommonKotlinModule.module,
             CommonNetworkModule.module,
-            CommonMonitoringModule(environment.apiUrl).module,
+            CommonMonitoringModule.module,
             SlackProtoModule(environment.apiUrl).module,
             SelfDestructAdapterModule.module,
             SelfDestructDomainModule.module,
