+++ self-destruct/service/src/jsMain/kotlin/com/gchristov/thecodinglove/selfdestruct/service/SelfDestructServiceModule.kt
@@ -5,7 +5,7 @@ import com.gchristov.thecodinglove.selfdestruct.domain.model.Environment
 import org.kodein.di.DI
 import org.kodein.di.bindSingleton
 
-class SelfDestructServiceModule(private val environment: Environment) : DiModule() {
+internal class SelfDestructServiceModule(private val environment: Environment) : DiModule() {
     override fun name() = "self-destruct-service"
 
     override fun bindDependencies(builder: DI.Builder) {
