+++ slack/service/src/jsMain/kotlin/com/gchristov/thecodinglove/slack/service/SlackService.kt
@@ -22,6 +22,7 @@ import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackInteractivityPubSub
 import com.gchristov.thecodinglove.slack.adapter.pubsub.SlackSlashCommandPubSubHandler
 import com.gchristov.thecodinglove.slack.domain.SlackDomainModule
 import com.gchristov.thecodinglove.slack.domain.model.Environment
+import com.gchristov.thecodinglove.slack.proto.SlackProtoModule
 
 suspend fun main() {
     // Ignore default Node arguments
@@ -48,12 +49,14 @@ private fun setupDi(environment: Environment): Either<Throwable, Unit> {
             CommonKotlinModule.module,
             CommonNetworkModule.module,
             CommonPubSubModule.module,
-            CommonMonitoringModule(environment.apiUrl).module,
+            CommonMonitoringModule.module,
             CommonFirebaseModule.module,
             SearchProtoModule(environment.apiUrl).module,
             SlackDomainModule.module,
             SlackAdapterModule(environment).module,
             SlackServiceModule(environment).module,
+            // The Slack service also uses the monitoring module so we need to link the protos specifically here.
+            SlackProtoModule(environment.apiUrl).module,
         )
     )
     return Either.Right(Unit)
