+++ slack/service/src/jsMain/kotlin/com/gchristov/thecodinglove/slack/service/SlackServiceModule.kt
@@ -5,7 +5,7 @@ import com.gchristov.thecodinglove.slack.domain.model.Environment
 import org.kodein.di.DI
 import org.kodein.di.bindSingleton
 
-class SlackServiceModule(private val environment: Environment) : DiModule() {
+internal class SlackServiceModule(private val environment: Environment) : DiModule() {
     override fun name() = "slack-service"
 
     override fun bindDependencies(builder: DI.Builder) {
